+++ src/panels/workflow/Workflow.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Panel, PanelHeader } from '@vkontakte/vkui';
 
-import Subsection from './subsection/Subsection';
+import Subsection from './Subsection/Subsection';
 
 const Workflow = ({ id }) => (
   <Panel id={id}>
